Added Group json option

+++ src/src/Commands/Group/GroupClearCommand.php
@@ -3,6 +3,7 @@
 namespace QIT_CLI\Commands\Group;
 
 use QIT_CLI\Cache;
+use QIT_CLI\TestGroup;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -13,8 +14,12 @@ class GroupClearCommand extends Command {
 	/** @var Cache */
 	protected $cache;
 
-	public function __construct( Cache $cache ) {
-		$this->cache = $cache;
+	/** @var TestGroup */
+	protected $test_group;
+
+	public function __construct( Cache $cache, TestGroup $test_group ) {
+		$this->cache      = $cache;
+		$this->test_group = $test_group;
 		parent::__construct();
 	}
 
@@ -24,8 +29,16 @@ class GroupClearCommand extends Command {
 	}
 
 	protected function execute( InputInterface $input, OutputInterface $output ) {
-		$this->cache->delete( 'group' );
+		$group = $this->test_group->get();
 
+		if ( empty( $group ) ) {
+			$output->writeln( '<error>No group found.</error>' );
+			return Command::FAILURE;
+		}
+
+		$this->test_group->output_group( $output );
+
+		$this->cache->delete( 'group' );
 		$output->writeln( '<info>Group cleared.</info>' );
 
 		return Command::SUCCESS;
